+++ Source/SwiftLintCore/Extensions/Request+SwiftLint.swift
@@ -2,15 +2,17 @@ import Foundation
 import SourceKittenFramework
 
 public extension Request {
-    static let disableSourceKit = {
+    nonisolated(unsafe) static var disableSourceKitOverride = false
+
+    static var disableSourceKit: Bool {
         #if SWIFTLINT_DISABLE_SOURCEKIT
         // Compile-time
         true
         #else
         // Runtime
-        ProcessInfo.processInfo.environment["SWIFTLINT_DISABLE_SOURCEKIT"] != nil
+        ProcessInfo.processInfo.environment["SWIFTLINT_DISABLE_SOURCEKIT"] != nil || disableSourceKitOverride
         #endif
-    }()
+    }
 
     func sendIfNotDisabled() throws -> [String: any SourceKitRepresentable] {
         // Skip safety checks if explicitly allowed (e.g., for testing or specific operations)
@@ -45,7 +47,7 @@ public extension Request {
         }
 
         guard !Self.disableSourceKit else {
-            queuedFatalError("SourceKit is disabled by `SWIFTLINT_DISABLE_SOURCEKIT`.")
+            queuedFatalError("SourceKit is disabled by configuration.")
         }
         return try send()
     }